+++ src/expressions/index.ts
@@ -6,3 +6,4 @@ export * as include from "./include";
 export * as loop from "./loop";
 export * as boolean_not from "./boolean_not";
 export * as arguments from "./arguments";
+export * as standard from "./standard";